requirements.txt
fastapi
uvicorn
openai
faiss-cpu
neo4j
requests
beautifulsoup4
python-dotenv
openai
langchain
langchain_openai
python-dotenv
tqdm
selenium